--- windows/dart_vlc_plugin.cpp
@@ -1,58 +1,0 @@
-/*
- * dart_vlc: A media playback library for Dart & Flutter. Based on libVLC & libVLC++.
- * 
- * Hitesh Kumar Saini
- * https://github.com/alexmercerind
- * saini123hitesh@gmail.com; alexmercerind@gmail.com
- * 
- * GNU Lesser General Public License v2.1
- */
-
-#include <flutter/method_channel.h>
-#include <flutter/plugin_registrar_windows.h>
-#include <flutter/standard_method_codec.h>
-#include "include/dart_vlc/dart_vlc_plugin.h"
-
-
-std::unique_ptr<flutter::MethodChannel<flutter::EncodableValue>> channel;
-
-namespace {
-
-
-    class DartVlcPlugin : public flutter::Plugin {
-    public:
-        static void RegisterWithRegistrar(flutter::PluginRegistrarWindows *registrar);
-
-        DartVlcPlugin();
-
-        virtual ~DartVlcPlugin();
-
-    private:
-        void HandleMethodCall(const flutter::MethodCall<flutter::EncodableValue> &method_call, std::unique_ptr<flutter::MethodResult<flutter::EncodableValue>> result);
-    };
-
-
-    void DartVlcPlugin::RegisterWithRegistrar(flutter::PluginRegistrarWindows *registrar) {
-        channel = std::make_unique<flutter::MethodChannel<flutter::EncodableValue>>(registrar->messenger(), "dart_vlc", &flutter::StandardMethodCodec::GetInstance());
-        auto plugin = std::make_unique<DartVlcPlugin>();
-        channel->SetMethodCallHandler(
-            [plugin_pointer = plugin.get()](const auto &call, auto result) {
-                plugin_pointer->HandleMethodCall(call, std::move(result));
-            }
-        );
-        registrar->AddPlugin(std::move(plugin));
-    }
-
-    DartVlcPlugin::DartVlcPlugin() {}
-
-    DartVlcPlugin::~DartVlcPlugin() {}
-
-    void DartVlcPlugin::HandleMethodCall(const flutter::MethodCall<flutter::EncodableValue> &methodCall, std::unique_ptr<flutter::MethodResult<flutter::EncodableValue>> result) {
-        /// No platform channel implementation after migration to FFI.
-        result->NotImplemented();
-    }
-}
-
-void DartVlcPluginRegisterWithRegistrar(FlutterDesktopPluginRegistrarRef registrar) {
-    DartVlcPlugin::RegisterWithRegistrar(flutter::PluginRegistrarManager::GetInstance()->GetRegistrar<flutter::PluginRegistrarWindows>(registrar));
-}
